+++ test/test_pattern.sh
@@ -743,7 +743,7 @@ expectfn "$clixon_cli -1f $cfg -l o set c threematch abcg" 255 '^CLI syntax erro
 if [ $BE -ne 0 ]; then
     new "Kill backend"
     # Check if premature kill
-    pid=$(pgrep -u $BUSER -f clixon_backend)
+    pid=$(pgrep -u root -f clixon_backend)
     if [ -z "$pid" ]; then
 	err "backend already dead"
     fi